+++ theme.js
@@ -24,5 +24,6 @@ function loadStyle(url) {
     loadScript("/appearance/themes/Dark+/script/module/timestamp.js");
     loadScript("/appearance/themes/Dark+/script/module/typewriter.js");
     
+    loadScript("/appearance/themes/Dark+/app/comment/index.js");
     // loadScript("/appearance/themes/Dark+/script/test/listener.js");
 })();